+++ CHANGES.md
@@ -27,6 +27,7 @@ Beta Releases
 * Added renderer support for `OES_element_index_uint`, which can improve performance by reducing batch sizes.
 * Added `czm_tangentToEyeSpaceMatrix` built-in GLSL function.
 * Added `WallGeometry` and `PolygonGeometry`.
+* Added `EllipseGeometry` and `CircleGeometry`.
 
 ### b17 - 2013-06-03